requirements.txt
# Danny AI Dependencies

# Strands Agent Framework
strands-agents>=1.27.0

# Core
anthropic>=0.18.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Logging and monitoring
structlog>=24.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# CLI
rich>=13.7.0
typer>=0.9.0

# Web API
fastapi>=0.100.0
uvicorn>=0.27.0

# AWS Services
boto3>=1.34.0

# Amazon Transcribe Streaming SDK
amazon-transcribe>=0.6.0
